+++ CHANGELOG.md
@@ -102,7 +102,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Unnecessary files are not included in release anymore. Added `files` section in `package.json`.
 
 ### Docs
-- Added extended example: [Plugin integration with asset management](examples/asset-management)
+- Added extended example: [Plugin integration with asset management](examples/cloud-front-service/asset-management)
 
 ## [3.0.0] - 2020-12-03
 
@@ -147,7 +147,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Breaking Changes
 - Ability to load template [configuration in any async way](https://github.com/Enase/serverless-ses-template/issues/15) (eg. from DB or API). \
-Template configuration file must return an async function now. Example see [here](examples/asset-management/ses-email-templates/index.js).
+Template configuration file must return an async function now. Example see [here](examples/cloud-front-service/asset-management/ses-email-templates/index.js).
 
 ### Added
 - List of supported regions updated according to [AWS RPS](https://aws.amazon.com/about-aws/global-infrastructure/regional-product-services/)